requirements.txt
asgiref==3.7.2
azure-core==1.27.0
azure-storage-blob==12.16.0
azure-storage-file-share==12.12.0
azure-storage-queue==12.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.1
django-dotenv==1.4.2
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
